fix float/long problems
in repl_insert_text and repl_erase_text

+++ sublimerepl.py
@@ -34,12 +34,12 @@ SETTINGS_FILE = 'SublimeREPL.sublime-settings'
 
 class ReplInsertTextCommand(sublime_plugin.TextCommand):
     def run(self, edit, pos, text):
-        self.view.insert(edit, pos, text)
+        self.view.insert(edit, long(pos), text)
 
 
 class ReplEraseTextCommand(sublime_plugin.TextCommand):
     def run(self, edit, start, end):
-        self.view.erase(edit, sublime.Region(start, end))
+        self.view.erase(edit, sublime.Region(long(start), long(end)))
 
 
 class Event:
